Move StandaloneXml, DomainXml, HostXml out of shared controller module
